Treat GOST like SIGALGS for cert verify
GOST can use a number of different digests for the handshake data in
processing a CertificateVerify. Therefore we should keep the handshake data
around until we have processed the CertificateVerify message so that we
can calculate the correct digest. This is much the same way that the
SIGALGS code works now.