Add the ability to have "_" in a cipher rule string
We are using "real" TLS1.3 names as per the RFC for the new ciphersuite.
These have an "_" character in them which we couldn't previously handle.

Also, when matching a ciphersuite if we are given an id, make sure we use it
otherwise we will match another ciphersuite which is identical except for
the TLS version.