Fix-up dh_test
The dh_test was failing because we now enforce a lower bound on the
modulus size that may be used. A number of locations in dhtest.c were
assuming that a very small modulus is valid.

A 512 bit lower bound was introduced by PR #9437 (commit 6de1fe908) and
subsequently amended by PR #9796 (commit feeb7ecd2f).

The CHANGES entry says this:

 * Enforce a minimum DH modulus size of 512 bits.