RFC 7919 FFDHE groups are now treated as DHX instead of DH in an EVP_PKEY
The RFC7919 FFDHE groups are all safe primes. That RFC specifies the groups
in terms of p, q and g (where q = (p-1)/2, because they are safe). However,
in 1.1.1, calling the function DH_new_by_nid() and specifying the nid for
one of these primes got you back a DH object with only p and g defined.
The q value was always NULL. In 3.0 we also provide the q value.

This makes a difference when we "assign" such a DH object to an EVP_PKEY.
During the assign we check whether the q value is present or not. If not
the key type for the EVP_PKEY is EVP_PKEY_DH. If it is then the key type
is EVP_PKEY_DHX.

This means that creating an RFC7919 DH object in 3.0 and assigning it to an
EVP_PKEY results in an EVP_PKEY with a different type to what you get with
1.1.1

The same applies even if you construct a DH object for these groups
"manually", i.e. by creating the DH object by DH_new(), and setting only
the p and g values via DH_set0_pqg(). In this case the DH_set0_pqg() is
sufficiently clever that it recognises that you have set a named group and
automatically fills in the q value.

This difference was causing a test failure in evp_extra_test.